implemented read and write for CGI communicaiton, separated struct for multiclient processing - functionnal for time out test now, simplified overall code

+++ srcs/MethodsCGI.cpp
@@ -2,7 +2,6 @@
 
 void	Methods::_launchCGI(void)
 {
-	// COUT << "Creating Pipes\n";
 	/* Pipe creation to communicate with the CGI program */
 	int pipefd_in[2];
 	int pipefd_out[2];
@@ -15,18 +14,14 @@ void	Methods::_launchCGI(void)
 	fcntl(pipefd_out[0], F_SETFL, O_NONBLOCK);
 	fcntl(pipefd_out[1], F_SETFL, O_NONBLOCK);
 
-	// COUT << "Forking\n";
 	/* Fork() the program for the CGI */
 	pid_t pid;
 	if ((pid = fork()) == -1)
 		serv->getParent()->abortServers("Error in fork()", strerror(errno));
-	/* Child program */
-	else if (pid == 0)
+	else if (pid == 0)	/* Child program */
 	{
 		close(pipefd_in[1]);
 		close(pipefd_out[0]);
-
-		// COUT << "Inside CHild\n";
 
 		/* Array for execve env */
 		_createEnvpMap();
@@ -34,7 +29,6 @@ void	Methods::_launchCGI(void)
 		if (envp == NULL)
 			exit(EXIT_FAILURE);
 
-		// COUT << "Creating Arguments\n";
 		/* Array for execve arguments */
 		_createArgvMap();
 		char ** argv = _createArgvArray();
@@ -47,7 +41,6 @@ void	Methods::_launchCGI(void)
 		// _displayArray(envp);
 		// _displayArray(argv);
 
-		// COUT << "Duping\n";
 		/* Duplicating Fd for STDIN and STDOUT */
 		if (dup2(pipefd_in[0], STDIN_FILENO) < 0 || close(pipefd_in[0])		/* Lecture par le CGI dans fd_in[0] */
 		|| dup2(pipefd_out[1], STDOUT_FILENO) < 0 || close(pipefd_out[1]))	/* Ecriture par le CGI dans fd_out[1] */
@@ -65,8 +58,7 @@ void	Methods::_launchCGI(void)
 		_freeArray(argv);
 		exit(EXIT_FAILURE);
 	}
-	/* Parent program */
-	else
+	else	/* Parent program */
 	{
 		// COUT << "Inside Parent\n";
 		close(pipefd_in[0]);
@@ -79,16 +71,13 @@ void	Methods::_launchCGI(void)
 
 void	Methods::_communicateWithCGI(int fd_in, int fd_out, pid_t pid)
 {
-	std::string receivedMessage;
 	bool	finishedWriting = 0;
 	bool	finishedReading = 0;
 	bool	CGIfinished = 0;
 
-	int status = 0;
 	Select	cgi;
 
-	/* Pas de timeout */
-	cgi.timeout.tv_sec = 0;
+	cgi.timeout.tv_sec = 1;
 	cgi.timeout.tv_usec = 0;
 
 	while (!(finishedWriting && finishedReading && CGIfinished))
@@ -116,6 +105,7 @@ void	Methods::_communicateWithCGI(int fd_in, int fd_out, pid_t pid)
 		{
 			case 0:
 			{
+				int status = 0;
 				if (waitpid(pid, &status, WNOHANG) == pid)
 				{
 					COUT << "Child was terminated";
@@ -138,17 +128,22 @@ void	Methods::_communicateWithCGI(int fd_in, int fd_out, pid_t pid)
 	/* STOP */	if (FD_ISSET(STDIN_FILENO, &cgi.readfds))	/* Keyboard was pressed, exiting server properly */
 				{
 					COUT << "Keyboard was pressed, killing CGI properly\n";
-					kill(pid, SIGKILL);
 					close(fd_in);
 					close(fd_out);
+					kill(pid, SIGKILL);
 					return ;
 				}
 	/* WRITE */	else if (FD_ISSET(fd_out, &cgi.writefds))
 				{
-					// COUT << "FD is available to write\n";
-					if (serv->resp.sendMsgCGI(fd_out, serv->req.body) == true)
+					if (_writeReqtoCGI(fd_out) == true)
 					{
-						// COUT << "Body to CGI|" << serv->req.body << "|\n";
+						if (client->clientClosed)
+						{
+							close(fd_out);
+							close(fd_in);
+							kill(pid, SIGKILL);
+							return ;
+						}
 						close(fd_out);
 						finishedWriting = 1;
 						COUT << "Sent EOF to CGI\n";
@@ -156,39 +151,17 @@ void	Methods::_communicateWithCGI(int fd_in, int fd_out, pid_t pid)
 				}
 	/* READ */	else if (FD_ISSET(fd_in, &cgi.readfds))
 				{
-					// COUT << "FD is available to read\n";
-					char	recv_buffer[MAX_HEADER_SIZE];
-
-					ssize_t receivedBytes = read(fd_in, &recv_buffer, MAX_HEADER_SIZE);
-					if (receivedBytes < 0)
+					if (_readCGItoResp(fd_in) == true)
 					{
-						close(fd_in);
-						serv->getParent()->abortServers("Error in read()", strerror(errno));
-					}
-					else if (receivedBytes == 0)
-					{
-						_parseCGIResponse(receivedMessage);
-
-						COUT << RED << "> CGI RESPONSE\n";
-						COUT << "Status: " << serv->resp.status_code << " " << serv->resp.reason_phrase << ENDL;
-						std::map<std::string, std::string, ci_less>::iterator begin = serv->resp.header_fields.begin();
-						while (begin != serv->resp.header_fields.end())
+						if (client->clientClosed)
 						{
-							COUT << begin->first + ": " + begin->second + "\r\n";
-							begin++;
+							close(fd_in);
+							kill(pid, SIGKILL);
+							return ;
 						}
-						COUT << "Body first line|" << serv->resp.body.substr(0, 100) << "|\n";
-						COUT << "Body Length |" << serv->resp.body.length() << "|\n";
-						COUT << RESET;
-
 						close(fd_in);
 						finishedReading = 1;
-					}
-					else
-					{
-						receivedMessage.append(recv_buffer, static_cast<size_t>(receivedBytes));
-						// COUT << "ReceivedBytes|" << receivedMessage << "|\n";
-						_parseCGIResponse(receivedMessage);
+						COUT << "Received EOF from CGI\n";
 					}
 				}
 			}
@@ -196,6 +169,39 @@ void	Methods::_communicateWithCGI(int fd_in, int fd_out, pid_t pid)
 	}
 	COUT << "CGI finished\n";
 	return ;
+}
+
+bool	Methods::_writeReqtoCGI(int & fd_out)
+{
+	ssize_t sentBytes = write(fd_out, &client->req.body.data()[_writtenBytes], client->req.body.length() - _writtenBytes);
+	if (sentBytes < 0)
+	{
+		if (sentBytes < 0)
+			CERR << "Error in write(): " << strerror(errno) << ENDL;
+		client->clientClosed = true;
+		return (true);
+	}
+	if ((_writtenBytes += static_cast<size_t>(sentBytes)) == client->req.body.length())
+		return (true);
+	return (false);
+}
+
+bool	Methods::_readCGItoResp(int & fd_in)
+{
+	char	recv_buffer[MAX_HEADER_SIZE];
+
+	ssize_t receivedBytes = read(fd_in, &recv_buffer, MAX_HEADER_SIZE);
+	if (receivedBytes <= 0)
+	{
+		if (receivedBytes == 0)
+			_parseCGIResponse();
+		_receivedMessage.clear();
+		_receivedMessage.reserve();
+		return (true);
+	}
+	_receivedMessage.append(recv_buffer, static_cast<size_t>(receivedBytes));
+	_parseCGIResponse();
+	return (false);
 }
 
 bool	Methods::_str_is(std::string str, int func(int))
@@ -211,64 +217,64 @@ bool	Methods::_str_is(std::string str, int func(int))
 	return (true);
 }
 
-bool	Methods::_parseHeaderField(std::string & receivedMessage)
+bool	Methods::_parseHeaderField(void)
 {
-	size_t size = receivedMessage.find("\r\n");
+	size_t size = _receivedMessage.find("\r\n");
 
 	if (size == 0)
 	{
-		receivedMessage.erase(0, 2);
+		_receivedMessage.erase(0, 2);
 		return (true);
 	}
 
 	size_t	pos = 0;
 	bool	return_value = 0;
 	size_t	osp = 0;
-	if ((pos = receivedMessage.find(":")) != std::string::npos)
+	if ((pos = _receivedMessage.find(":")) != std::string::npos)
 	{
-		if (receivedMessage.find(": ", pos) == pos)
+		if (_receivedMessage.find(": ", pos) == pos || _receivedMessage.find(":\t", pos) == pos)
 			osp = 1;
 
-		std::string key = receivedMessage.substr(0, pos);
+		std::string key = _receivedMessage.substr(0, pos);
 
-		if (serv->req.strFindCaseinsensitive(key, "Status") == 0)
+		if (client->req.strFindCaseinsensitive(key, "Status") == 0)
 		{
-			serv->resp.status_code = receivedMessage.substr(7 + osp, 3);
-			serv->resp.reason_phrase = receivedMessage.substr(11 + osp, size - 11 - osp);
-			if (!_str_is(serv->resp.status_code, isdigit) || !_str_is(serv->resp.reason_phrase, isprint))
+			client->resp.status_code = _receivedMessage.substr(7 + osp, 3);
+			client->resp.reason_phrase = _receivedMessage.substr(11 + osp, size - 11 - osp);
+			if (!_str_is(client->resp.status_code, isdigit) || !_str_is(client->resp.reason_phrase, isprint))
 				return_value = true; // Status value is incorrect - return true to finish parsing
 		}
 		else
 		{
-			if (serv->req.strFindCaseinsensitive(key, "Location") == 0) // a implementer davantage
+			if (client->req.strFindCaseinsensitive(key, "Location") == 0) // a implementer davantage
 			{
-				serv->resp.status_code = "301";
-				serv->resp.reason_phrase = "Found";
+				client->resp.status_code = "301";
+				client->resp.reason_phrase = "Found";
 			}
-			else if (serv->req.strFindCaseinsensitive(key, "Content-Type") == 0)
+			else if (client->req.strFindCaseinsensitive(key, "Content-Type") == 0)
 			{
-				serv->resp.status_code = "200";
-				serv->resp.reason_phrase = "OK";
+				client->resp.status_code = "200";
+				client->resp.reason_phrase = "OK";
 			}
-			serv->resp.header_fields.insert(std::make_pair(key, receivedMessage.substr(pos + 1 + osp, size - pos - 1 - osp)));
+			client->resp.header_fields.insert(std::make_pair(key, _receivedMessage.substr(pos + 1 + osp, size - pos - 1 - osp)));
 		}
 	}
-	receivedMessage.erase(0, size + 2);
+	_receivedMessage.erase(0, size + 2);
 	return (return_value);
 }
 
-void	Methods::_parseCGIResponse(std::string & receivedMessage)
+void	Methods::_parseCGIResponse(void)
 {
-	size_t size = receivedMessage.size();
+	size_t size = _receivedMessage.size();
 	static bool	HeaderIncomplete = true;
 
-	while (HeaderIncomplete && receivedMessage.find("\r\n") != std::string::npos)
-		HeaderIncomplete = !_parseHeaderField(receivedMessage);
+	while (HeaderIncomplete && _receivedMessage.find("\r\n") != std::string::npos)
+		HeaderIncomplete = !_parseHeaderField();
 
-	if (!receivedMessage.empty())
+	if (!_receivedMessage.empty())
 	{
-		serv->resp.body += receivedMessage;
-		receivedMessage.clear();
+		client->resp.body += _receivedMessage;
+		_receivedMessage.clear();
 	}
 	else if (!size)
 		HeaderIncomplete = true;
@@ -281,30 +287,30 @@ void	Methods::_createEnvpMap(void)
 // AUTH_TYPE
 	/* _checkAuthenticate already assign the correct value */
 // CONTENT_LENGTH
-	if (serv->req.headers.find("Content-Length") != serv->req.headers.end())
-		_envp["CONTENT_LENGTH"] = serv->req.headers.find("Content-Length")->second;
+	if (client->req.headers.find("Content-Length") != client->req.headers.end())
+		_envp["CONTENT_LENGTH"] = client->req.headers.find("Content-Length")->second;
 	else
 		_envp["CONTENT_LENGTH"] = "";
 // CONTENT_TYPE
-	if (serv->req.headers.find("Content-Type") != serv->req.headers.end())
-		_envp["CONTENT_TYPE"] = serv->req.headers.find("Content-Type")->second;
+	if (client->req.headers.find("Content-Type") != client->req.headers.end())
+		_envp["CONTENT_TYPE"] = client->req.headers.find("Content-Type")->second;
 	else
 		_envp["CONTENT_TYPE"] = "";
 // GATEWAY_INTERFACE
 	_envp["GATEWAY_INTERFACE"] = "CGI/1.1";
 // PATH_INFO
-	_envp["PATH_INFO"] = serv->req.uri;	// SUFFIXE DE L'URI UNIQUEMENT?
+	_envp["PATH_INFO"] = client->req.uri;	// SUFFIXE DE L'URI UNIQUEMENT?
 // PATH_TRANSLATED
 	_envp["PATH_TRANSLATED"] = _path;
 // QUERY_STRING
 	_envp["QUERY_STRING"] = _query;	/* put the search identifier in the uri if any (query-string part of the uri) */
 // REMOTE_ADDR
-	tmp << inet_ntoa(serv->req.client->address.sin_addr);
+	tmp << inet_ntoa(client->socket.address.sin_addr);
 	_envp["REMOTE_ADDR"] = tmp.str();	/* Get client IP adress */
 	tmp.str("");
 // REMOTE_IDENT
 	// A VERIFIER
-	tmp << ntohs(serv->serv_port.address.sin_port) << ", " << ntohs(serv->req.client->address.sin_port);
+	tmp << ntohs(serv->serv_port.address.sin_port) << ", " << ntohs(client->socket.address.sin_port);
 	if (_envp["REMOTE_USER"].empty())
 		tmp << " : ERROR : HIDDEN-USER";	//  6195, 23 : ERROR : NO-USER
 	else
@@ -314,9 +320,9 @@ void	Methods::_createEnvpMap(void)
 // REMOTE_USER
 	/* _checkAuthenticate already assign the correct value */
 // REQUEST_METHOD
-	_envp["REQUEST_METHOD"] = serv->req.method;
+	_envp["REQUEST_METHOD"] = client->req.method;
 // REQUEST_URI
-	_envp["REQUEST_URI"] = serv->req.uri; /* Check with PATH_INFO */
+	_envp["REQUEST_URI"] = client->req.uri; /* Check with PATH_INFO */
 // SCRIPT_NAME
 	_envp["SCRIPT_NAME"] = _cgi_path;
 // SERVER_NAME
@@ -335,13 +341,13 @@ void	Methods::_createEnvpMap(void)
 	_envp["SERVER_SOFTWARE"] = "HuntGaming/1.0";
 
 // ADDITIONAL IMPLEMENTATION-DEFINED CGI HEADER FIELDS
-	for (std::map<std::string, std::string, ci_less>::iterator it = serv->req.headers.begin(); it != serv->req.headers.end(); ++it)
+	for (std::map<std::string, std::string, ci_less>::iterator it = client->req.headers.begin(); it != client->req.headers.end(); ++it)
 	{
-		if (serv->req.strFindCaseinsensitive(it->first.substr(0, 2), "X-") != std::string::npos)
+		if (client->req.strFindCaseinsensitive(it->first.substr(0, 2), "X-") != std::string::npos)
 		{
 			std::string result = "HTTP_" + it->first;
-			result = serv->req.transform(result, toupper);
-			result = serv->req.transform(result, serv->req.tounderscore);
+			result = client->req.transform(result, toupper);
+			result = client->req.transform(result, client->req.tounderscore);
 			_envp[result] = it->second;
 		}
 	}
@@ -353,7 +359,7 @@ char **	Methods::_createEnvpArray(void)
 {
 	size_t	envp_size = 0;
 
-	for (; serv->getParent()->envp[envp_size]; ++envp_size);
+	for (; serv->getParent()->getEnvp()[envp_size]; ++envp_size);
 	char ** array = static_cast<char **>(malloc(sizeof(char *) * (_envp.size() + envp_size + 1)));
 	if (array == NULL)
 	{
@@ -383,9 +389,9 @@ char **	Methods::_createEnvpArray(void)
 		++i;
 		++begin;
 	}
-	for (int j = 0; serv->getParent()->envp[j]; ++j)
+	for (int j = 0; serv->getParent()->getEnvp()[j]; ++j)
 	{
-		array[i] = strdup(serv->getParent()->envp[j]);
+		array[i] = strdup(serv->getParent()->getEnvp()[j]);
 		if (array[i] == NULL)
 		{
 			CERR << "Error in dup(): " << strerror(errno) << ENDL;
